+++ authentication/gradle.properties
@@ -1,4 +1,3 @@
-
 #
 # Copyright (C) 2024. Uber Technologies
 #
